// src/objects/enums.ts
export enum NodeGroupType {
  Sources = "sources",
  Classification = "classification",
  Processors = "processors",
  DateTime = "datetime",
  Deeplearning = "deeplearning",
  Heuristics = "heuristics",
  Machinelearning = "machinelearning",
  Modelintegration = "modelintegration",
  Optimization = "optimization",
  Timeseries = "timeseries",
  Sinks = "sinks"
}